Create generator subpackage and start moving files
