+++ lib/src/generator/generator_frontend.dart
@@ -4,7 +4,7 @@
 
 import 'dart:async';
 
-import 'package:dartdoc/src/generator.dart';
+import 'package:dartdoc/src/generator/generator.dart';
 import 'package:dartdoc/src/logging.dart';
 import 'package:dartdoc/src/model/model.dart';
 import 'package:dartdoc/src/model_utils.dart';
