+++ lib/src/generator/generator_utils.dart
